[R] Refactored Sample models
